mergt shear.py into auto_augment.py
